+++ src/Config/DefaultConfig.php
@@ -26,7 +26,6 @@ class DefaultConfig extends Config
     {
         parent::__construct(null, array(
             self::PULI_DIR => '.puli',
-            self::INSTALL_FILE => '{$puli-dir}/install-file.json',
             self::DUMP_DIR => '{$puli-dir}/dump',
             self::WRITE_REPO => '{$puli-dir}/resource-repository.php',
             self::READ_REPO => '{$write-repo}',