+++ tests/Config/ConfigFile/ConfigFileManagerTest.php
@@ -96,12 +96,12 @@ class ConfigFileManagerTest extends \PHPUnit_Framework_TestCase
                 $config = $configFile->getConfig();
 
                 \PHPUnit_Framework_Assert::assertSame('my-puli-dir', $config->get(Config::PULI_DIR));
-                \PHPUnit_Framework_Assert::assertSame('my-puli-dir/my-install-file.json', $config->get(Config::INSTALL_FILE));
+                \PHPUnit_Framework_Assert::assertSame('my-puli-dir/my-dump', $config->get(Config::DUMP_DIR));
             }));
 
         $this->manager->setConfigKeys(array(
             Config::PULI_DIR => 'my-puli-dir',
-            Config::INSTALL_FILE => '{$puli-dir}/my-install-file.json',
+            Config::DUMP_DIR => '{$puli-dir}/my-dump',
         ));
     }
 
@@ -117,26 +117,26 @@ class ConfigFileManagerTest extends \PHPUnit_Framework_TestCase
     public function testGetConfigKeyReturnsRawValue()
     {
         $this->configFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->configFile->getConfig()->set(Config::INSTALL_FILE, '{$puli-dir}/install-file.json');
+        $this->configFile->getConfig()->set(Config::DUMP_DIR, '{$puli-dir}/my-dump');
 
-        $this->assertSame('{$puli-dir}/install-file.json', $this->manager->getConfigKey(Config::INSTALL_FILE));
+        $this->assertSame('{$puli-dir}/my-dump', $this->manager->getConfigKey(Config::DUMP_DIR));
     }
 
     public function testGetConfigKeys()
     {
         $this->configFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->configFile->getConfig()->set(Config::INSTALL_FILE, '{$puli-dir}/install-file.json');
+        $this->configFile->getConfig()->set(Config::DUMP_DIR, '{$puli-dir}/my-dump');
 
         $this->assertSame(array(
             Config::PULI_DIR => 'my-puli-dir',
-            Config::INSTALL_FILE => '{$puli-dir}/install-file.json',
+            Config::DUMP_DIR => '{$puli-dir}/my-dump',
         ), $this->manager->getConfigKeys());
     }
 
     public function testRemoveConfigKey()
     {
         $this->configFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->configFile->getConfig()->set(Config::INSTALL_FILE, 'install-file.json');
+        $this->configFile->getConfig()->set(Config::DUMP_DIR, 'my-dump');
 
         $this->configFileStorage->expects($this->once())
             ->method('saveConfigFile')
@@ -145,7 +145,7 @@ class ConfigFileManagerTest extends \PHPUnit_Framework_TestCase
                 $config = $configFile->getConfig();
 
                 \PHPUnit_Framework_Assert::assertNull($config->get(Config::PULI_DIR, false));
-                \PHPUnit_Framework_Assert::assertSame('install-file.json', $config->get(Config::INSTALL_FILE, false));
+                \PHPUnit_Framework_Assert::assertSame('my-dump', $config->get(Config::DUMP_DIR, false));
             }));
 
         $this->manager->removeConfigKey(Config::PULI_DIR);
@@ -154,7 +154,7 @@ class ConfigFileManagerTest extends \PHPUnit_Framework_TestCase
     public function testRemoveConfigKeys()
     {
         $this->configFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->configFile->getConfig()->set(Config::INSTALL_FILE, 'install-file.json');
+        $this->configFile->getConfig()->set(Config::DUMP_DIR, 'my-dump');
 
         $this->configFileStorage->expects($this->once())
             ->method('saveConfigFile')
@@ -163,9 +163,9 @@ class ConfigFileManagerTest extends \PHPUnit_Framework_TestCase
                 $config = $configFile->getConfig();
 
                 \PHPUnit_Framework_Assert::assertNull($config->get(Config::PULI_DIR, false));
-                \PHPUnit_Framework_Assert::assertNull($config->get(Config::INSTALL_FILE, false));
+                \PHPUnit_Framework_Assert::assertNull($config->get(Config::DUMP_DIR, false));
             }));
 
-        $this->manager->removeConfigKeys(array(Config::PULI_DIR, Config::INSTALL_FILE));
+        $this->manager->removeConfigKeys(array(Config::PULI_DIR, Config::DUMP_DIR));
     }
 }