+++ tests/Config/ConfigFile/Writer/ConfigJsonWriterTest.php
@@ -51,7 +51,6 @@ class ConfigJsonWriterTest extends JsonWriterTestCase
         $configFile = new ConfigFile();
         $configFile->getConfig()->merge(array(
             Config::PULI_DIR => 'puli-dir',
-            Config::INSTALL_FILE => '{$puli-dir}/my-install-file.json',
             Config::DUMP_DIR => '{$puli-dir}/my-repo',
             Config::READ_REPO => '{$puli-dir}/my-repository.php',
             Config::WRITE_REPO => '{$puli-dir}/my-repository-dump.php',